can't delete the last domain

+++ backend/services/domains.py
@@ -38,4 +38,6 @@ class UserDomainsService:
         ).model_dump()
     
     def delete_domain(self,user_id: int, domain_id: str):
+        if self.domain_persistence.count_domain(user_id) == 1:
+            raise HTTPException(status_code=409, detail={'status': 'LAST_DOMAIN'})
         return self.domain_persistence.delete_domain(user_id, domain_id)
